e2e-pytest/pyproject.toml
[project]
name = "e2e-pytest"
version = "0.1.0"
description = "Python pytest end-to-end test suite"
requires-python = ">=3.12"
dependencies = []

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-playwright>=0.5",
    "httpx>=0.27",
    "asyncpg>=0.30",
    "docker>=7.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
